+++ lib/config.go
@@ -17,6 +17,7 @@ var (
 		"delete":             true,
 		"save_attachments":   true,
 		"remove_attachments": true,
+		"export_mailbox":     true,
 	}
 )
 
@@ -156,7 +157,12 @@ func (r Rule) RemoveAttachments() bool {
 	return strings.Contains(r.Actions, "remove_attachments")
 }
 
-// SaveAttachments returns whether a rule is set to delete messages
+// SaveAttachments returns whether a rule is set to save attachments
 func (r Rule) SaveAttachments() bool {
 	return strings.Contains(r.Actions, "save_attachments")
+}
+
+// ExportMailbox returns whether a rule is set to export matching messages to mbox
+func (r Rule) ExportMailbox() bool {
+	return strings.Contains(r.Actions, "export_mailbox")
 }